imapclient: move VANISHED handling into vanished.go
Split handleVanished out of expunge.go, where it was unrelated to the
EXPUNGE command, into a dedicated vanished.go pairing with
vanished_test.go. Also clarify the UnilateralDataHandler.Vanished doc
comment: note the response is unsolicited and describe the earlier flag
in terms of what it indicates.
